Changing the default neighbor "skin" distance so that you get good energy conservation using the AMH-Go potential. Note that this may effect efficiency, and you might want to turn it back to a lower value if you using other potentials that use a neighbor list but you are not using AMH-Go.
I am also changing the default time step to 2 fs because we have seen some problems with energy conservation when using a time step of 5 fs. Again, this will effect efficiency and some combination of potentials may allow for larger time steps to be taken. However, I think that the default time step should be on the conservative side.